Alekhya | MOBN-510 | Extract common method to helper class
